Clean up and added new styles
Fixed style issues so far. Footer is hoverble for desktop mode and can scroll to the end to see the footer in mobiles. Navbar styles are added and has responsive design for mobiles. Removed sidebar component because its no longer needed in the application.  Navbar toggling much better than earlire on mobiles.